The POM file was updated and is creating jar file successfully, just type "mvn clear install" on root directory to create this jar.
Next steps is remove this container structure that isn't necessary and
go deep on consumer and producer sides.

offtopic (related with mas2ssp project): It was the first test of this
elementar component inside of mas2ssp project and it is running! In
mas2ssp compiled by jedit is able to receive opcda data. Many libs must
be add in lib folder of mas2ssp.